Install libffi everytime we install pipenv
which is currently all the time.

We expect a future story to only install pipenv when it is needed, so
this code is structured to make that adjustment easy by calling
installFfi inside installPipEnv, and keeping the call to HandleFfi and
its knowledge of which app dependencies should cause libffi to be
installed.

Hopefully, the future change should only involve adding an if-statment
around the contents of the InstallPipEnv function and adjusting the unit
tests.

[#156016877]

Signed-off-by: Andrew Meyer <ameyer@pivotal.io>