Use v3-push for app update tests
Using CF CLI v6 is only natively compatible with a CAPI version that contains a bug.
This bug causes secondary pushes to fail sometimes. For now, use the v3-push command which
uses a later CAPI version. This is a temporary fix. To fix this properly, we need to upgrade
Cutlass to use CF CLI later than v6, preferably v8.