Don't fall back to unvendored installs
Doing so is an anti-feature. In general this is the tool saying
"I know you have explicitly requested to do X, but now that this
has failed I will do the exact opposite of what you requested".
In general, when a user explicitly requests X and this fails
the correct course of action is to provide an error message.

Specifically:

* When vendored installs fail, unvendored installs will often also fail, but
  may fail for different reasons. Example: build dependencies may not be
  available in the buildpack. This sends the user off on a wild
  goose chase (fixing the second problem, rather than the original one)

* When unvendored installs will in fact not fail (i.e. the intention
  of the now-removed code) the user's expectation that whatever is in the
  "vendor" directory is what's actually installed is violated. This violates
  reproducability of installs, and opens a can of worms security-wise
  (e.g. the vendored packages may have been scanned previously, whereas
  arbitrary downloads are generally not).

History: The anti-feature was introduced in 518f933f15de, without providing
reasoning why this is desirable.